chore(ui): Delete computed parameters from scenario execution + Move KeyValue model to dataset
